add initial services in Startup, move enums to Enum folder
