+++ backend/PetSupport.API/Data/ApplicationDbContext.cs
@@ -1,0 +1,6 @@
+﻿namespace PetSupport.API
+{
+    internal class ApplicationDbContext
+    {
+    }
+}
