+++ backend/PetSupport.Core/Entities/Petsitter.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
+using PetSupport.Core.Entities;
 
 namespace PetSupport.Data.Entities
 {
@@ -6,8 +8,8 @@ namespace PetSupport.Data.Entities
     {
        public string Description { get; set; }
         public string Password { get; set; }
-        public string Photo { get; set; }
+        public string PhotoId { get; set; }
         public string City { get; set; }
-        public List<Service> Services { get; set; }
+        public IQueryable<Service> Services { get; set; }
     }
 }
