+++ backend/PetSupport.Core/Entities/User.cs
@@ -12,6 +12,7 @@ namespace PetSupport.Data.Entities
         public string Surname { get; set; }
 
         public string Email { get; set; }
+
         public string PhoneNumber { get; set; }
     }
 }
