+++ backend/PetSupport.Core/Enums/ServiceType.cs
@@ -1,0 +1,11 @@
+﻿namespace PetSupport.Core.Enums
+{
+    public enum ServiceType
+    {
+        Boarding,
+        HouseSitting,
+        DogWalking,
+        DropInVisit,
+        DoggyDayCare
+    }
+}
