+++ backend/PetSupport.Core/Enums/Unit.cs
@@ -1,0 +1,9 @@
+﻿namespace PetSupport.Core.Enums
+{
+    public enum Unit
+    {
+        Day,
+        Hour,
+        Night
+    }
+}
