+++ backend/PetSupport.Infrastructure.Data/PetSupport.Infrastructure.Data.csproj
@@ -9,4 +9,10 @@
     <Folder Include="Seed\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational.Design" Version="1.1.6" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.0" />
+  </ItemGroup>
+
 </Project>
